Revert "feat(backend): rename `db.py` to `sqlite.py`"
This reverts commit 88bee96ca3876e545293c9d12dbd19e466386b03.
