+++ invokeai/app/services/board_image_record_storage.py
@@ -4,7 +4,7 @@ from abc import ABC, abstractmethod
 from typing import Optional, cast
 
 from invokeai.app.services.models.image_record import ImageRecord, deserialize_image_record
-from invokeai.app.services.shared.sqlite import SqliteDatabase
+from invokeai.app.services.shared.db import SqliteDatabase
 from invokeai.app.services.shared.pagination import OffsetPaginatedResults
